requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.2.0
bidict==0.23.1
certifi==2024.6.2
chainlit==0.7.700
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.5.14
Deprecated==1.2.14
distro==1.9.0
exceptiongroup==1.2.1
fastapi==0.100.1
fastapi-socketio==0.0.10
filetype==1.2.0
frozenlist==1.4.1
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.64.1
grpcio-tools==1.62.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.7
importlib_metadata==7.1.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.9
langchain-openai==0.1.8
langchain-qdrant==0.1.1
langchain-text-splitters==0.2.1
langsmith==0.1.81
Lazify==0.4.0
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.35.3
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
orjson==3.10.5
packaging==23.2
portalocker==2.8.2
protobuf==4.25.3
pydantic==2.7.4
pydantic_core==2.18.4
PyJWT==2.8.0
PyMuPDF==1.24.5
pymupdf4llm==0.0.5
PyMuPDFb==1.24.3
python-dotenv==1.0.1
python-engineio==4.9.1
python-graphql-client==0.4.3
python-multipart==0.0.6
python-socketio==5.11.3
PyYAML==6.0.1
qdrant-client==1.9.1
regex==2024.5.15
requests==2.32.3
simple-websocket==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.27.0
syncer==2.0.3
tenacity==8.4.1
tiktoken==0.7.0
tomli==2.0.1
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
uptrace==1.24.0
urllib3==2.2.2
uvicorn==0.23.2
watchfiles==0.20.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.19.2